+++ core/test/com/google/inject/AllTests.java
@@ -157,10 +157,10 @@ public class AllTests {
     suite.addTestSuite(Jsr330Test.class);
 
     // multibindings tests
-    suite.addTestSuite(com.google.inject.multibindings.MapBinderTest.class);
-    suite.addTestSuite(com.google.inject.multibindings.MultibinderTest.class);
-    suite.addTestSuite(com.google.inject.multibindings.OptionalBinderTest.class);
-    suite.addTestSuite(com.google.inject.multibindings.RealElementTest.class);
+    suite.addTestSuite(com.google.inject.internal.MapBinderTest.class);
+    suite.addTestSuite(com.google.inject.internal.MultibinderTest.class);
+    suite.addTestSuite(com.google.inject.internal.OptionalBinderTest.class);
+    suite.addTestSuite(com.google.inject.internal.RealElementTest.class);
     suite.addTestSuite(com.google.inject.multibindings.ProvidesIntoTest.class);
 
     return SuiteUtils.removeSuppressedTests(suite, SUPPRESSED_TEST_NAMES);
